+++ src/main/java/com/greedy/mokkoji/api/user/dto/request/kakao/KakaoSocialLoginRequest.java
@@ -1,4 +1,4 @@
-package com.greedy.mokkoji.api.user.dto.request;
+package com.greedy.mokkoji.api.user.dto.request.kakao;
 
 public record KakaoSocialLoginRequest(
         String code
