+++ src/main/java/com/greedy/mokkoji/api/user/dto/resopnse/kakao/KakaoAccessTokenResponse.java
@@ -1,11 +1,10 @@
-package com.greedy.mokkoji.api.user.dto.resopnse;
+package com.greedy.mokkoji.api.user.dto.resopnse.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoProfileResponse(
-        String nickname,
-        boolean isDefaultNickname
+public record KakaoAccessTokenResponse(
+        String accessToken
 ) {
 }
