+++ src/main/java/com/greedy/mokkoji/api/user/dto/resopnse/kakao/KakaoAccountResponse.java
@@ -1,12 +1,10 @@
-package com.greedy.mokkoji.api.user.dto.resopnse;
+package com.greedy.mokkoji.api.user.dto.resopnse.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserInfoResponse(
-        String id,
-        KakaoAccountResponse kakaoAccount
+public record KakaoAccountResponse (
+        KakaoProfileResponse profile
 ) {
 }
-
