+++ src/main/java/com/greedy/mokkoji/api/user/dto/resopnse/kakao/KakaoUserInfoResponse.java
@@ -1,10 +1,12 @@
-package com.greedy.mokkoji.api.user.dto.resopnse;
+package com.greedy.mokkoji.api.user.dto.resopnse.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoAccountResponse (
-        KakaoProfileResponse profile
+public record KakaoUserInfoResponse(
+        String id,
+        KakaoAccountResponse kakaoAccount
 ) {
 }
+
