+++ src/main/java/com/greedy/mokkoji/api/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.greedy.mokkoji.api.user.service;
 
 import com.greedy.mokkoji.api.jwt.JwtUtil;
-import com.greedy.mokkoji.api.user.dto.resopnse.KakaoUserInfoResponse;
+import com.greedy.mokkoji.api.user.dto.resopnse.kakao.KakaoUserInfoResponse;
 import com.greedy.mokkoji.common.exception.MokkojiException;
 import com.greedy.mokkoji.db.user.entity.User;
 import com.greedy.mokkoji.db.user.repository.UserRepository;
